Removed Python 2.5 dependent calls (rpartition) Forced task output file to absolute path Stopped setup script from changing working directory Added documentation and added error handling to setup script Added more standard BONIC daemons into config during setup Fixed spacing bug in old pymw_setup script

+++ pymw/pymw/interfaces/boinc_setup.py
@@ -91,6 +91,12 @@ def get_winworker_path():
     return path
 
 def install_pymw(project_path):
+    """Installs a default app named "pymw" into BOINC and
+    sets up the assimilator to look in the current directory
+    to pick up workunits. In addition to setting up the 
+    application, it will also check to see if the project
+    is running, it not, it will attempt to start it.
+    """
     logging.debug("---------------------------------")
     logging.debug("Installing PyMW BOINC application")
     logging.debug("---------------------------------")
@@ -108,6 +114,8 @@ def install_pymw(project_path):
     logging.debug("---------------------")
 
 def check_daemons(project_path):
+    """Checks for STOP_TRIGGER, if present calls bin/start    
+    """
     stopped = os.path.exists(os.path.join(project_path, STOP_TRIGGER))
     
     if stopped:
@@ -119,26 +127,56 @@ def check_daemons(project_path):
         logging.debug("BOINC daemon status: RUNNING")
 
 def setup_config(task_path):
-    logging.info("writing pymw_assimilator in config.xml")
+    """Adds appropriate daemons to the BOINC config.xml file.
+    """
+    logging.info("Adding PyMW Assimilator daemon to config.xml")
     config = configxml.ConfigFile().read()
     
-    # Remove old instances of pymw_assimilator
-    for daemon in config.daemons:
-        if daemon.cmd[0:16] == 'pymw_assimilator':
-            config.daemons.remove_node(daemon)
+    ## Remove old instances of pymw_assimilator
+    #for daemon in config.daemons:
+    #    if daemon.cmd[0:16] == 'pymw_assimilator':
+    #        config.daemons.remove_node(daemon)
     
     # Append new instance of pymw_assimilator to config.xml
-    config.daemons.make_node_and_append("daemon").cmd = "pymw_assimilator.py -d 3 -app pymw -pymw_dir " + task_path
-    config.write()
+    asm = "pymw_assimilator.py -d 3 -app pymw -pymw_dir " + task_path
+    add_daemon(config, asm, "pymw_assimilator")
+
+    # add a file deleter, ignoring batches
+    asm = "file_deleter -d 3 -ignore_batches"
+    add_daemon(config, asm, "file_deleter")
+    
+    # add a default validator
+    asm = "sample_trivial_validator -d 3 -app pymw"
+    add_daemon(config, asm, "sample_trivial_validator")
+
+    # add a feeder and a transitioner
+    asm = "feeder -d 3"
+    add_daemon(config, asm, "feeder")
+    asm = "transitioner -d 3"
+    add_daemon(config, asm, "transitioner")
+    
+    #config.daemons.make_node_and_append("daemon").cmd = asm
+    #config.write()
     return config
 
+def add_daemon(config, command, remove):
+    # first try to remove it
+    for daemon in [d for d in config.daemons if d.cmd.startswith(remove)]:
+        config.daemons.remove_node(daemon)
+        logging.debug("Removing existing daemon: %s" % daemon.cmd)
+
+    # now add in the command
+    config.daemons.make_node_and_append("daemon").cmd = command
+    logging.debug("Appending daemon: %s" % command)
+    config.write()
+    
 def setup_project():
     # Append new instance of pymw worker to project.xml
     project = projectxml.ProjectFile().read()
     found = False
     for element in project.elements:
         if element.name == 'pymw':
-            logging.info("PyMW client applications are already installed")
+            logging.info("PyMW client application is already present in project.xml")
             return
     
     project.elements.make_node_and_append("app").name = "pymw"
@@ -152,48 +190,91 @@ def setup_project():
 def install_apps(config):
     # Install worker applications
     app_dir = os.path.join(config.config.app_dir, "pymw")
-    if os.path.isdir(app_dir):
-        logging.info("PyMW client applications are already installed")
-        return
-
-    os.mkdir(app_dir)
+    if not os.path.isdir(app_dir):
+        os.mkdir(app_dir)
+    
     install_posix(app_dir, i686_pc_linux_gnu, POSIX_WORKER, "Linux")
     install_posix(app_dir, i686_apple_darwin, POSIX_WORKER, "Apple")
     install_windows(app_dir, windows_intelx86)
     
     # Call update_versions
     project_home = config.config.app_dir.rpartition('/')[0]
-    os.chdir(project_home)
-    os.system("xadd")
-    os.system("update_versions --force --sign")
-        
+    
+    os.system("cd %s; xadd" % project_home)
+    os.system("cd %s; update_versions --force --sign" % project_home)
+
+def file_exists(path, name, data=None):
+    """Checks to see if a file exists, if so, prints a message if
+    the Name parameter is not None. 
+    
+    If the data parameter is not None, the file is opened and the
+    data parameter is written to it. Return True if the file exists, 
+    False otherwise.
+    """
+    if os.path.exists(path):
+        if name != None:
+            logging.debug("%s already installed, skipping file")
+        return True
+    if data == None: return False
+    f = open(path, "w")
+    try: f.writelines(data)
+    finally: f.close()
+    return False
+  
 def install_windows(app_dir, app_name):
+    """Installs the windows application by copying:
+     - [Windows-worker].exe
+     - [Windows-worker].exe.file_ref_info
+     - pymw_app.py
+     - pymw_app.py.file_ref_info
+    into the apps directory of the BOINC project
+    """
     # windows_intelx86
     logging.info("setting up client application for Windows platform")
     win_dir = os.path.join(app_dir, windows_intelx86)
     win_exe = os.path.join(win_dir, windows_intelx86)
+    win_exe_ref = os.path.join(win_dir, windows_intelx86 + ".file_ref_info")
+    pymw_app = os.path.join(win_dir, "pymw_app.py")
+    pymw_app_ref = os.path.join(win_dir, "pymw_app.py.file_ref_info")
+    
     workerpath = get_winworker_path()
     if not workerpath or not os.path.exists(workerpath): 
         logging.critical("Unable to locate the windows worker executable, windows clients will be disabled")
         logging.critical("The path returned was: " + workerpath)
         return
-    os.mkdir(win_dir)
-    shutil.copy(workerpath, win_exe)
-    f = open(os.path.join(win_dir, windows_intelx86 + ".file_ref_info"), "w").writelines(FILE_REF)
-    open(os.path.join(win_dir, "pymw_app.py"), "w").writelines(PYMW_APP)
-    open(os.path.join(win_dir, "pymw_app.py.file_ref_info"), "w").writelines(FILE_REF)
+    
+    if not os.path.exists(win_dir):os.mkdir(win_dir)
+    if not file_exists(win_exe, "Windows main-app"):
+        shutil.copy(workerpath, win_exe)
+    
+    file_exists(win_exe_ref, None, FILE_REF)
+    file_exists(pymw_app, None, PYMW_APP)
+    file_exists(pymw_app_ref, None, FILE_REF)
+    
     os.chmod(win_exe, stat.S_IRWXU | stat.S_IRGRP | stat.S_IROTH)
-    logging.info("client application for Windows platform set up successfully")
+    logging.info("Client application for Windows platform set up successfully")
     
 def install_posix(app_dir, app_name, worker, friendly_name):
+    """Installs a generic posix shell script by coping:
+     - the-script
+     - the-script.file_ref_info
+     - pymw_app.py
+     - pymw_app.py.file_ref_info
+    into the apps directory of the BOINC project. 
+    """
     logging.info("setting up client application for " + friendly_name + " platform")
     target_dir = os.path.join(app_dir, app_name)
     target_exe = os.path.join(target_dir, app_name)
-    os.mkdir(target_dir)
-    open(target_exe, "w").writelines(worker)
-    open(os.path.join(target_dir, app_name + ".file_ref_info"), "w").writelines(FILE_REF)
-    open(os.path.join(target_dir, "pymw_app.py"), "w").writelines(PYMW_APP)
-    open(os.path.join(target_dir, "pymw_app.py.file_ref_info"), "w").writelines(FILE_REF)
+    target_exe_ref = os.path.join(target_dir, app_name + ".file_ref_info")
+    pymw_app = os.path.join(target_dir, "pymw_app.py")
+    pymw_app_ref = os.path.join(target_dir, "pymw_app.py.file_ref_info")
+
+    if not os.path.exists(target_dir): os.mkdir(target_dir)
+    file_exists(target_exe, friendly_name, worker)
+    file_exists(target_exe_ref, None, FILE_REF)
+    file_exists(pymw_app, None, PYMW_APP)
+    file_exists(pymw_app_ref, None, FILE_REF)
+
     os.chmod(target_exe, stat.S_IRWXU | stat.S_IRGRP | stat.S_IROTH)
     logging.info("client application for " + friendly_name + " platform set up successfully")
